requirements-dev.txt

# Dev lint code
flake8
# Dev test code
pytest-cov
# Dev coverage
codecov
